Add files via upload
I created a website with 3 examples of uploading videos onto webpages. Embedding, Embedding live video, and uploading raw HTML. The first problem I ran into was I tried running HTML on p5 and after realizing I needed a program like Sublime I still didnt switch the extension to HTML. After a quick tutorial from Kailen I was able to figure out how to apply all the notes I took down on video to web output and use some of the coding. With uploading the raw HTML i took a random video I had and moved it to a folder that I could find and used the name of the file in my code so that the website could retrieve the information from my computer. Then I added a player with code I found online, muted the video, and looped it. For the Youtube video it was as simple as copy and pasting the Embed Code. Lastly, for embedding live video I had to download other third party streaming interface. I used XSplitBroadcaster and Youtube simultaneously and took the embed code and put it into my website. I also did